Finish refactoring ibc-relayer-cosmos
