+++ crates/relayer-cosmos/src/base/mod.rs
@@ -1,5 +1,2 @@
 pub mod all_for_one;
-pub mod error;
-pub mod impls;
-pub mod traits;
 pub mod types;
