--- crates/relayer-cosmos/src/contexts/full/mod.rs
@@ -1,4 +1,0 @@
-pub mod birelay;
-pub mod builder;
-pub mod chain;
-pub mod relay;
