+++ crates/relayer-cosmos/src/contexts/mod.rs
@@ -1,2 +1,5 @@
-pub mod full;
-pub mod min;
+pub mod birelay;
+pub mod builder;
+pub mod chain;
+pub mod relay;
+pub mod transaction;
