+++ crates/relayer-cosmos/src/full/impls/relay.rs
@@ -11,27 +11,4 @@ impl<Relay> OfaFullRelay for CosmosRelayWrapper<Relay>
 where
     Relay: CosmosFullRelay,
 {
-    fn is_retryable_error(_: &Error) -> bool {
-        false
-    }
-
-    fn max_retry_exceeded_error(e: Error) -> Error {
-        e
-    }
-
-    async fn should_relay_packet(&self, packet: &Self::Packet) -> Result<bool, Self::Error> {
-        Ok(self
-            .relay
-            .packet_filter()
-            .channel_policy
-            .is_allowed(&packet.source_port, &packet.source_channel))
-    }
-
-    fn src_chain_message_batch_sender(&self) -> &CosmosBatchSender {
-        self.relay.src_chain_message_batch_sender()
-    }
-
-    fn dst_chain_message_batch_sender(&self) -> &CosmosBatchSender {
-        self.relay.dst_chain_message_batch_sender()
-    }
 }
