--- crates/relayer-cosmos/src/full/types/builder.rs
@@ -1,9 +1,0 @@
-pub struct CosmosFullBuilderWrapper<Builder> {
-    pub builder: Builder,
-}
-
-impl<Builder> CosmosFullBuilderWrapper<Builder> {
-    pub fn new(builder: Builder) -> Self {
-        Self { builder }
-    }
-}
