+++ crates/relayer-cosmos/src/types/batch.rs
@@ -3,8 +3,8 @@ use tendermint::abci::Event as AbciEvent;
 use tokio::sync::mpsc::{UnboundedReceiver, UnboundedSender};
 use tokio::sync::oneshot::Sender as SenderOnce;
 
-use crate::base::error::Error;
-use crate::base::types::message::CosmosIbcMessage;
+use crate::types::error::Error;
+use crate::types::message::CosmosIbcMessage;
 
 pub type CosmosBatchPayload = (
     Vec<CosmosIbcMessage>,
