+++ crates/relayer-cosmos/src/types/mod.rs
@@ -1,2 +1,4 @@
 pub mod batch;
+pub mod error;
+pub mod message;
 pub mod telemetry;
